HDDcoin v1.2.3 Update
Updated HDDcoin to v1.2.3

+++ install-gui.sh
@@ -25,20 +25,20 @@ if [ "$(uname)" = "Linux" ]; then
 		# Debian/Ubuntu
 		UBUNTU=true
 		sudo apt-get install -y npm nodejs libxss1
-	elif type yum &&  [ ! -f "/etc/redhat-release" ] && [ ! -f "/etc/centos-release" ] && [ ! -f /etc/rocky-release ]; then
+	elif type yum &&  [ ! -f "/etc/redhat-release" ] && [ ! -f "/etc/centos-release" ] && [ ! -f /etc/rocky-release ] && [ ! -f /etc/fedora-release ]; then
 		# AMZN 2
 		echo "Installing on Amazon Linux 2."
 		curl -sL https://rpm.nodesource.com/setup_12.x | sudo bash -
 		sudo yum install -y nodejs
-	elif type yum && [ ! -f /etc/rocky-release ] && [ -f /etc/redhat-release ] || [ -f /etc/centos-release ]; then
+	elif type yum && [ ! -f /etc/rocky-release ] && [ ! -f /etc/fedora-release ] && [ -f /etc/redhat-release ] || [ -f /etc/centos-release ]; then
 		# CentOS or Redhat
 		echo "Installing on CentOS/Redhat."
 		curl -sL https://rpm.nodesource.com/setup_12.x | sudo bash -
 		sudo yum install -y nodejs
-	elif type yum && [ -f /etc/rocky-release ]; then
+	elif type yum && [ -f /etc/rocky-release ] || [ -f /etc/fedora-release ]; then
                 # RockyLinux
-                echo "Installing on RockyLinux"
-                dnf module enable nodejs:12
+                echo "Installing on RockyLinux/Fedora"
+                sudo dnf module enable nodejs:12
                 sudo dnf install -y nodejs
         fi
 
@@ -95,6 +95,7 @@ if [ ! "$CI" ]; then
 	npm install
 	npm audit fix || true
 	npm run build
+	python ../installhelper.py
 else
 	echo "Skipping node.js in install.sh on MacOS ci."
 fi
